add register user domain model
