+++ src/model/mod.rs
@@ -1,4 +1,5 @@
 pub mod auth_middleware;
 pub mod login_response;
+pub mod register_user;
 pub mod token;
 pub mod user;
